MB-60382: Fix intermittent GetAllKeys Failure
We should wait till all upserts have been persisted before initiating a
GetAllKeys command. The test was seen to fail when a document is still
pending and not yet persisted. AllKeysCallback skips durability-prepared
keys, and on machines where the flusher is slow, those keys are not
returned.

This patch, retireves the last seqno and persistest all seqnos up to and
including this one, ensuring that no key remains in the prepare state.

Change-Id: I479480f5c56444658b3d89f3f7423bf00bd96a6d
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/211195
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Mohammad Zaeem <mohammad.zaeem@couchbase.com>