kv_stats_log does not handle non-bucket commands correctly
stats.log contains two basic types of blocks.

For commands associated with a bucket, we use * separators.

  ================
  description
  command
  ================
  ****************
  bucket-1

   stat-1: ....

For commands not associated with a bucket, we do not have *.

  ===============
  description
  command
  ===============
  output

kv_stats_log did not handle the second case correctly, as it was always
expecting a bucket name.

In addition to fixing that, I've also made sure that we don't print
duplicate entries when using the list option, which was possible for
mcstat outputs due to how the output is separated.

Change-Id: I2b2ab9b40c5053987ca527312723cd6a96f89e1b
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/211517
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Vesko Karaganev <vesko.karaganev@couchbase.com>