MB-61385: Refine CAS correction for Replication/front-end
KVBucket:set/deleteWithMeta share code paths with DCP replication. This
meant that as poisoned documents are replicated by DCP, the consumer
will pass the documents through the CAS bound checks and inadvertently
remove data over a rebalance.

To prevent this unexpected behaviour, a new invalidCasStrategy has been
added, specifically for DCP replication, and the respective strategy is
implemented for the type of request.

Change-Id: I74d4c2934e5201f228fd68de0ed7aa0c212db28f
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/211797
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>