MB-62516: FlushAccounting::StatisticsUpdate account for ModifyCollection
The flushedItem boolean member of StatisticsUpdate should be updated 
when an item is flushed that should be erased should a collection be 
dropped in the same flush batch. ModifyCollection is not being 
accounted for but does not fundamentally break anything as the flusher
falls back to an manifest lock and check.

This commit tidies the code, accounts for ModifyCollection and adds a 
new test covering the exact sequence which was originally of concern 
flush of modify/drop.

Change-Id: Id9175cefa02d67b637a0bbf39847d2abc5ab2dcb
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/211940
Tested-by: Jim Walker <jim@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>