MB-43068: Add CommitStrategy which allows fallback to non-majority
Added CommitStrategy as a property of the ADM, which is set to
MajorityAck (Commit requires ack from majority > 1).

The second option is MajorityAckFallbackToMasterAckOnly which is being
implemented as part of this work. This strategy allows durable writes
when the replication topology does not have enough nodes to ack for
majority > 1. This exludes null topology (before the topology is set)
and of course the case of 3 replicas which we do not support.

The motivation here is to allow SyncWrites to be committed only on the
active when all replicas have been failed over or when no replicas are
configured (these two cases are treated the same).

This patch only enables the SyncWrites to be created, but does not
commit them yet.

Change-Id: I7d6324ae9559b664e74161f00752057babb1f0b0
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/212340
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>