MB-43068: Commit SyncWrites only on active when nodes < majority
SyncWrites can be satisfied with status ToCommitNotDurable when the
replication chain does not have enough replicas for majority ack.

In the case of two replication chains, both need to be satisfied to the
SyncWrite to be committed. If one chain has size=1 and the other size=2,
the SyncWrite will be acked on the second chain before commit.

Change-Id: I8579e34ed94fc8106034b8efddeb90f780017fab
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/212341
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>