MB-62596: Update getNumItemsForCursor to be constant time
Change the CheckpointManager/Cursor code so they work together
to compute items available.

CheckpointManager tracks "totalItems" which is a monotonically
increasing counter of items put into Checkpoints, e.g. every
mutation,  checkpoint_start/end and queueVBState, i.e. anything
which should be accounted in getNumItemsForCursor (so not the
queue_op::empty item).

This can be thought of as a number-line where totalItems is the
end of the line.

Cursor now tracks it's position on this number-line and we can
obtain items remaining by subtracting the position from the
totalItems.

Change-Id: I6d85bd951e272b0003a81fbc9e3afa170f5dcbd5
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/212288
Tested-by: Jim Walker <jim@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>