MB-41351: Rename many collection functions/types
Create/Flush of a collection is now better thought of as "begin".

Drop/Delete of a collection can thus be "end".

Rename many related functions/types using begin instead of names
containing create/flush.

This commit is less stringent around removing the "drop" language
as it does sometimes work better as a collection does not just end
it ends and disappears.

Change-Id: I04f22aa2b7165aeae3e0ec6cbc8ee867b2337ea0
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/212395
Tested-by: Jim Walker <jim@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>