MB-62665: Generate enum definitions for enum configuration keys
Changes the codegen script to generate enum classes.

These enums have the same name as the configuration key, so bucket_type
has a BucketType enum. The values are derived from the enum validator
specified in configuration.json. They are generated within the
cb::config namespace.

Functions to parse the strings into the enums and back are also
generated. String formatting functions format_as and to_string are
supplied for all enums.

To avoid creating a single patch which refactors a lot of
configuration-related code to use enums, I've kept the current
std::string-returning getters, but suffixed then with "String".
The enum-returning getters do not have a suffix.

This results in:

  enum class cb::config::BucketType { Persistent, Ephemeral };

  std::string Configuration::getBucketTypeString() const;
  cb::config::BucketType Configuration::getBucketType() const;
  std::string to_string(cb::config::BucketType);
  std::string_view format_as(cb::config::BucketType);

Changed the existing code to call the get*String variants. Refactoring
patches will follow.

Change-Id: Ic9bf955f59dd0b6711061675fa6aa04338bae30e
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/212475
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Vesko Karaganev <vesko.karaganev@couchbase.com>