MB-27800: Avoid rollback if no purge has run between a connection bounce
- If the remotePurgeSeqno included in the STREAM_REQ is the same as the
localPurgeSeqno for the vbucket, a rollback can be avoided.

- Introduce a new connection open flag:
DcpOpenFlag::SendSnapshotMarkerV2_2 - which indicates if the purgeSeqno
should be included in the Snapshot Marker responses.

- If a connection is opened with the above flag set, all stream requests
should include their view of the purgeSeqno (if one was received in
the previous Snapshot marker, 0 if this is the first time a stream
request is being initiated). Parse the new STREAM_REQ packet and pass it
down to the failover-table specific functions.

Change-Id: I4ae28d6904a81b6d24aa45bd526d8a185d00214a
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/213029
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Hareen Kancharla <hareen.kancharla@couchbase.com>