MB-41351: Track a collection-flush through the flusher and erase
Detect the flush as it is persisted and update a variety of collection
related structures so that item/disk accounting reflects the flush.

Flush stat accounting generally re-uses the existing drop-collection
logic.

1) When a flush(cid) is seen in the flusher-batch, track the highest seqno
of all flush(cid).

2) "ignore" item counting for the persistence of any flushed item(cid)
which has a seqno before the flush(cid) detected in 1). This ensures
the collection's stats only account for items that came after the flush.

3) The droppedCollections persisted metadata is also updated to track
the span of the flushedCollection - allowing a later purge to know what
to remove. The span is tracked as 0 to the flush(cid) detected in 1).

4) Finally collection purging is triggered just like drop-collection.

This commit only supports couchstore - which is the most straightfoward
backend to get to work.

Two tests are added to cover the flush and purge.

1) A basic test with no items (which is easier to debug)
2) A test which flushes with items and where the keys are chosen so that
those updates which are after the flush must be seen as inserts.

Change-Id: I62267b8957410df3823f235e1ae0afddb66f85b0
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/212573
Tested-by: Jim Walker <jim@couchbase.com>
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>