MB-62722: Don't reschedule callback with full send buffer
Don't trigger a callback if the send buffer towards the client
exceeds the threshold unless the connection is used for DCP
traffic.

Without the fix there was an edge case where clients could starve
other clients by sending a (large) pipeline of commands to the
server causing it to eventually exceed the maximum limit and
when it happens the connection would "constantly" be retriggered
just to be backed out again.

Change-Id: I572439f7a0751fdbc9ee2536c7c73b5c50ab415f
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/212707
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>