[testapp] Temporary bump bucket memory to 300MB
There are a number of unit tests which use the old pattern
of:

    connection->sendCommand();
    connection->recvResponse();

and don't deal with tmpfail and out of memory with retry.

When using the magma buckets we hit the memory issue
earlier causing sporadic test failures.

Increase the bucket limit to 300MB until we refactor the
tests use "execute" which would retry

Change-Id: I81c622fca91ebc0b6b01fcae7d779c466c12a461
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/213242
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>