[Refactor] Don't keep parsed results in Response class
Simplify the logic with the Response packets to always
parse and build up the parsed data when the caller
tries to access the data in the response packet.

This simplifies the logic in the packet code as the
classes don't need to overload both an "assign" method
(used to store the raw data read off the socket) in
addition to the move constructor (used to call
execute) and have both methods wrap into a method
to parse the data to keep it in the cached copy.

Most use of these functions will only try to fetch
the result from the packet _once_ anyway which
means this extra caching have little value

Change-Id: I720ae2454dcb8c31d93f92e7c221f615f7383ad7
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/213291
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>