MB-62847: Fix DCP premature SeqnoAdvance when only set_vb_state is queued
lastReadSeqno monotonic exception occurred when a single set_vb_state
item was processed after a disk/memory merged snapshot. The stream
state was able to trigger a SeqnoAdvance in-place of the skipped
set_vb_state, setting lastReadSeqno to the seqno of the set_vb_state.
The next actual mutation shares the same seqno and triggers the
lastReadSeqno monotonic exception.

The bug occurs partly because of the value curChkSeqno carries after
cursor registration from the backfill scheduling, in the unit test
it would be assigned 2, which at that time is the set_vb_state.
As curChkSeqno is then used as input to isSeqnoGapAtEndOfSnapshot
it compares equally against lastSentEndSnapshot, which is 2 because
of the extended snapshot.

Fix by tracking the seqnos of all replication items pulled from
checkpoint and use that instead of curChkSeqno as input to
isSeqnoGapAtEndOfSnapshot.

Change-Id: If9f211982a9824a461e2a1b52e75869d43241cfc
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/213361
Tested-by: Jim Walker <jim@couchbase.com>
Well-Formed: Restriction Checker
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>