MB-62814: Allow 1 vBucket to utilise more than 1 BGFetcher
Use as many BGFetchers as Reader threads. Before, we defaulted to the
number of shards instead.

Also, when the number of BGFetchers is more than the number of vBuckets,
assign vBuckets more than 1 BGFetcher. When a vBucket has more than 1
BGFetcher, use the temp_iniital SV's CAS as a key to choose the
BGFetcher to use.

In production, we use NumShards = NumCPUs by default, and NumReaders
depends on configuration. When NumShards < NumReaders, we're not
utilising all Reader threads, now we will be. When NumShards >
NumReaders, we've had more task scheduling, but the Reader pool
utilisation should be 100% before and after this change.

When BGFetchers > vBuckets, not all BGFetchers were used. Now all
BGFetchers are used, but if BGFetchers % vBuckets != 0, the remaining
BGFetchers are unequally distributed across vBuckets and some vBuckets
can have an additional BGFetcher. Assuming uniform workload, that
should be okay.

Change-Id: Icc5554cb47468d656af36accb080de6d49e1397d
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/213112
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Vesko Karaganev <vesko.karaganev@couchbase.com>