MB-62940: Collection stream purgeSeqno logic for Ephemeral
This patch implements the optimisation which allows us to skip the
purgeSeqno (MB-32358) for ephememral buckets.

Without this patch, streams against ephemeral buckets can be created in
the case covered by MB-32358, but will immediately be closed with status
Rollback as soon as the backfill runs, because the backfill code did not
implement the same logic.

The reason we didn't catch this earlier is that the relevant tests did
not run for ephemeral buckets. I've changed that and had to fix a few
tests in the progress to call the generic purgeTombstones function.

Change-Id: I3aea446bc2274af25dff44bac3b45945922f3b92
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/213528
Tested-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Well-Formed: Restriction Checker