MB-62961: Account new items after checkpoint expel
SuccessNewItem vs SuccessPersistAgain only decide if
CheckpointManager::queueDirty should increment the numItems
(critically for this MB increment totalItems).

In the 'de-dup' case where the original has been expelled return
SuccessNewItem. In terms of accounting this is a new item.

An existing test-case hits this condition and is updated to
check both numItems and the 'items-remaining' for the cursor.

Change-Id: Ia16076274c985866b5faacc839a91732f482ed8a
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/214270
Tested-by: Jim Walker <jim@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>