MB-63159: Use the correct purgeSeqno in DCPBackfillMemoryBuffered
Read the purgeSeqno from the Sequence list, instead of the "stat"
variant from VBucket, which may be not be the latest value.

Before this fix, a backfill could miss deletes.

I also had to refactor the bufferedMemoryBackfillPurgeGreaterThanStart,
as before, it _relied_ on the fact that the backfill could check the
VBucket purgeSeqno and that it can be trivially set to any value.

Change-Id: I87adc53b68dd9c3d18b6e91c01f79dcb6d7b0c0e
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/214401
Tested-by: Build Bot <build@couchbase.com>
Well-Formed: Restriction Checker
Reviewed-by: Mohammad Zaeem <mohammad.zaeem@couchbase.com>