MB-62856: Add support to return xattrs with RangeScan
Add new root level key that enables RangeScan to return xattrs with each
document. This prevents having to do a subdoc_multi_lookup to retrieve
xattrs for each document, which can be costly for a large amount.

System xattrs will be removed if the user does not have the privilege.

This is supported for both range-scan and random-sample configurations.

Change-Id: I655a01a11f857c5bceea69457ae03ecb2def2e37
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/214408
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Faizan Alam <faizan.alam@couchbase.com>