MB-62703: End backfilling (disk) streams which make no progress
A DCP stream which is backfilling can indefinitely hold disk resources
if it hangs. This has been seen to happen and even though the connection
owning the stream responds to DCP_NOOP messages. The server must check
for this condition and end the stream if no progress is made.

ScanContext is given an "abstract" Position type which allows the
DCP backfill class to monitor progress of the scan and check the
time. No change of position within the configured time results the
stream being ended with reason "Slow", releasing the disk resources.

A new config parameter "dcp_backfill_idle_limit_seconds" controls how
long the server will tolerate a backfill with no progress. This
parameter can be changed, but any new value only takes effect with
scans started after the parameter is changed. The default is 120
seconds which should be ample time to determine a client is faulty
without causing too much damage.

Ephemeral buckets are addressed in the next commit.

Change-Id: I23d050999ac601ff301da4fa3c64f9c75cc3c0ab
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/214579
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Jim Walker <jim@couchbase.com>