MB-63258: Use unbuffered io for reading from --stdin
As part of process startup memcached may get extra configuration
via commands sent over stdin (to make sure they're not visible
as part of the process environment or written to disk).

This mechanism used fgets() read input from stdin for simplicity
and portability, but it came in conflict with our "stdin check"
used later on.

Given that buffered IO may read multiple chunks of data from
the kernel and put in the userspace buffer, we could end up
reading out commands intended for our "stdin check", but we
would only try to read data there when we got a notification
from the kernel that we had new data to read.

We've previously had number of problems with the "stdin check"
logic on linux and rewritten the logic a few times (caused
by deadlocks in libc trying to acquire a shared mutex etc), and
we don't want to change the logic in there.

To work around the problem this patch use raw io to read
character by character from stdin as part of reading extra
configuration commands.

Change-Id: I304b54653b13a8b9cd4ba3ec3e9a141e6118356c
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/214820
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>