[BP] MB-63341: Expel set-vb-state from checkpoints
Three commits were created for MB-63341 and one further fix was
later found.

This backport commit brings togther all required changes into one
commit for neo. It is made up of two of the three MB-63341 commits
and the single commit created for MB-64246.

Included from MB-63341 are:

https://review.couchbase.org/c/kv_engine/+/215273 - improved debug
https://review.couchbase.org/c/kv_engine/+/215274 - the expel fix

Not included from MB-63341:

https://review.couchbase.org/c/kv_engine/+/215573

This is because for neo, the test that was adjusted in that commit
does not require the change, the cursor in neo is in the correct
place.

Included from MB-64246:

https://review.couchbase.org/c/kv_engine/+/222523

This is a fix for a latent bug introduced by the earlier MB-63341
changes.

Change-Id: I53e889bb6d88fc07ff7d8df65a0e14d4ee8e65d2
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/228121
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Jim Walker <jim@couchbase.com>
Well-Formed: Restriction Checker