MB-63380: Remove support for "stats reset"
Return "not supported" where we previously tried to reset a handful
of values. The primary motivation for removing the ability to reset
statistics is that it isn't very well defined which values we reset
or don't reset, and when only some values may have been reset it
may lead to confusion later on (why is X so low when Y indicates
something else etc).

The functionality is of limited value in production.

Change-Id: Ib23ae9b33fb4b5e5fcd27d5d159b4e913d4764d3
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/215410
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>