MB-54274: Don't delete temp item immediately
Current behavior:

- A temp item is inserted in the HT if a GET is requested for a deleted
  key in a full-eviction bucket.
- A bgfetch is scheduled & temp item is replaced with the actually
  delete value from disk when the bgfetch completes.
- When the front-end thread retries the GET, the temp-item is
  immediately deleted.

While this makes sense - it's sub-optimal when we have concurrent GET's
for the same deleted key.

Consider the following scenario:

1. Client-1 requests (GET) a deleted item in a full-eviction bucket.
2. We create a temp-item (mark it temp_init) & queue a bg-fetch.
3. Client-2 requests the same deleted item.
4. We notice the temp-item in the hash-table & queue another bg-fetch.
5. The bg-fetch completes, the front-end cookie for Client-1 is notified.
6. A Get it issued again & the temp item is deleted.
7. When the bg-fetch for client-2 completes it notices the temp-item is
   deleted & another bg-fetch is scheduled.

In the worst case we would have scheduled 3 bgfetches which could have
been easily served with a single bgfetch.

An important question though is when & how will this temp item be
eventually removed?

We currently have a expiry pager scheduled every 10 mins & that will
eventually clean-up the temp-items.

In addition to the above cleanup, we have an additional guardrail to
prevent unlimited growth of the temp items. A new configuration knob
"ht_temp_items_allowed_percent" (default = 10%) is added which limits
the number of temp items allowed in the hashtable as a percentage of the
HT size. The number of temp items allowed is updated on each HT when the
next HT resizer task runs.

Note, it can happen so at a later point some items are deleted & the
temp items percentage could be greater than
"ht_temp_items_allowed_percent".

Change-Id: I945eb7945b5fd869ed1ed973f5c16b6f5293f9ab
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/214421
Reviewed-by: Dan Owen <owend@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>