MB-61736: Make BgfetchSucceedsUntilMutationWatermark deterministic
Previously, this test set the bucket quota based on the size of the
stored item's value. This setup ensured that after the first item was
fetched, the second item would fail due to exceeding the mutation
watermark. However, this test would fail intermittently on CV because
the memory usage after the first item was BgFetched did not always
reach the mutation watermark.

To make this test deterministic, I have introduced a testing hook before
creating an item after it is read from disk. Using this hook, we adjust
the bucket quota relative to the current memory usage, ensuring that the
current usage becomes 93% of the bucket quota (mutation watermark). This
adjustment is made before creating the second item so that it will fail
to be BgFetched.

Change-Id: If0fe7d30a84fc10217614fa5afbddb58a9203f53
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/215891
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>