MB-62178: Add support for migrating off keys in magma
Pass the list of obsolete keys down to magma as part of
compaction.

NOTE: The implementation in magma may not be intuitive to
the caller as when a list of keys is provided it use a
different API in magma which *only* considers any segments
using one of the keys in the provided.

Change-Id: Iac9ce4db0aee5c6f753cfcf56e48c04c213f3bc3
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/216047
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>