MB-62250: Prepare to tell Magma to start/stop backup
This patch adds the logic for calling Magma::StartBackup and
Magma::StopBackup when a vbucket changes to active.

This is implemented by calling the appropriate function after a vbstate
flush, to ensure that newly started backup will see the vbucket in the
expected state.

This design requires a vbstate flush for the parameter change to
'continuous_backup_enabled' to take effect.

In the following patch, a vbstate will be queued when the parameter
changes.

Change-Id: I5827cde23a9d3153f997437000da40ea4285f14f
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/216184
Tested-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>