MB-62250: Start/StopBackup on continuous_backup_enabled change
The Start/StopBackup logic is triggered in the flusher path after
vbstate change. When the bucket parameter is changed at runtime, we need
to queue a vbstate to trigger that logic.

In the previous change, I'd forgotten to account for the
needsToBePersisted() optimisation in snapshotVBucket. In this patch,
I've split the updateCachedVBState() from the postVBStateFlush() and
will now call postVBStateFlush() on the early return path, even when
we don't end up writing to disk.

This results in the backup being correctly started/stopped whenever the
parameter changes.

Change-Id: Ic5b89cae0b26b462b7270b9fbb8399941403796a
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/216559
Tested-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>