MB-63201: Add GetEx and GetExReplica commands
These extended versions of the command returns the full value
including xattrs. The returned data will only contain the
system xattrs if the user holds the SystemXattrRead privilege.

Given that the system would like to do as little work as possible
the client must enable JSON, XATTR and SnappyEverywhere feature
in order to use the new command.

Change-Id: I2d22858e2aee42b89ea8fff345eb318e6455f7de
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/216740
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Trond Norbye <trond.norbye@couchbase.com>