MB-63618: Clear DcpConnHandler* from Connection on disconnect
The use of a raw pointer in Connection exposes a use-after-free
issue that is observed when get_failover_log uses the async pattern
(when warmup causes it to block).

This fix deliberately is non-invasive (avoiding a change to ownership)
and ensures the Connection raw pointer is cleared when we're on
the pre-destruct ConnHandler path. The pointer is cleared outside
of the ~ConnHandler as destruction of the object can occur on another
task.

Connection though exposes a isDCP method, which uses the pointer
value to provide a boolean return. When we clear the pointer on the
disconnect path, the frontend still may invoke scheduleDcpStep and
expects isDCP to remain returning true. This commit adds a separate
boolean to ensure isDCP remains true even after clearing the pointer.

Change-Id: Ibb0c7a15d58c286c91162a5476a6d5ab1c97f279
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/216805
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Jim Walker <jim@couchbase.com>