Refactor: Remove "boilerplate" code for adding new commands
Previously adding a new command required updating a bunch of functions
by adding another "case" statement inside the switch for opcodes. This
patch change the logic to have a fixed array with an element for each
opcode where one set the various "attributes" for the command (which
means the boilierplate code for adding a new command gets reduced to
populating the entry in the array).

Change-Id: I74f868a8fc137d965c5a47328b6be4b2b69e6f7d
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/216904
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>