MB-63431: Skip bloomfilter check for DCP mutations
When a dcp mutation is processed on the replica, we want to avoid the
expensive magma bloomfilter check (It was noticed in perf tests it takes
2x the time to process each mutation with the additional bloomfilter
check).

From code inspection the passive stream code always invokes setWithMeta
with GenerateBySeqno::No & all front-end ops are invoked with
GenerateBySeqno::Yes.

Use genBySeqno to distinguish the two operations & avoid the bloomfilter
check while processing a DCP mutation.

Change-Id: Idd546bcef4aa78010b1ff8d53cc99585db8d823a
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/216860
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>