MB-63766: Use target_link_libraries for ep tests
By using target_link_libraries as referencing the test
libraries (static and object) instead of using the
old $<TARGET_OBJECTS:libname> we no longer need to
add all dependencies for the library to all targets
where it is used.

In addition to that remove the special "ep_objs" library as we
no longer use a shared object for ep which means that ep and
ep_objs are identical.

Change-Id: Ic53b8ce16b34d4c53cecbd149fe55d67084f69a5
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/217150
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>