MB-63765: Check op CAS and locked status for deleteWithMeta
Fixes the special case where a DeleteWithMeta without a value finds a
value with xattrs. In that case, the existing system xattrs need
preserving. On that code path, we did not check the following:

 - that the request CAS matches
 - that the document is unlocked (or the request CAS is the same)

I've refactored the check into a method checkCasForWrite and used it in
a few other places -- the behaviour for other commands does not change
as they already had the correct checks.

Unfortunately the fix to deleteWithMeta revealed some broken tests.

 - The test test_delete_with_meta_nonexistent was checking if a
   deleteWithMeta can "delete" an nonExistent item using the "correct"
   cas. It used getMeta to "read" the cas of the nonExistent item. Since
   the item does not exist, it the call returned no_such_key and cas=0
   always.
 - An STItemPager test used cas=-1 in the deleteWithMeta, which is
   incorrect but never failed because we didn't validate the cas.

Change-Id: Ibd1c1e25115c75fe858c75aba26cd7d56610b56e
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/217207
Tested-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>
Well-Formed: Restriction Checker