MB-27800: Recheck purgeSeqno hasn't changed before backfill runs
We currently check if the remotePurgeSeqno (purge seqno included in the
last snapshot marker sent to the client) is the same as the purgeSeqno
on a vbucket before accepting a streamRequest. Perform the same check
again during the backfill to avoid any unnecessary rollback.

Also includes minor fixes described below:

- When we removed snapshot marker version 2.1 (supporting PiTR) we
  dropped the size of the extras buffer allocated for encoding a
  snapshot marker to support only SnapshotMarkerV2_0. Which breaks the
  buffer size checks when encoding SnapshotMarkerV2_2. Fixed that too
  via this patch.

- convert purge seqno to network byte order while encoding the snapshot
  marker.

Change-Id: I9185f8be90ec66d781e908aedab3d039aa3db2aa
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/217547
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Hareen Kancharla <hareen.kancharla@couchbase.com>