MB-63850: Remove CheckpointMemRecoveryTask periodic wake-up
When CheckpointMemRecoveryTask is triggered periodically every 5
seconds, it has the effect of freeing memory anytime the lower_mark is
exceeded. The periodic execution is not necessary since the task also
wakes up when inserting into CM and the upper_mark being exceeded.
This also addresses the behaviour of recovering memory anytime the
lower_mark has been exceeded, allowing the upper_mark to be reached.

Change-Id: Ifbf5abceb33b42fa12367d8ed26b3d045c3db71c
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/217939
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>