MB-64182: Toggle history eviction on/off when continuous backup is on
We need to prevent history eviction during replay of the WAL when
continuous backup is on. This is because the WAL replay can cause
history to be evicted before the continuous backup has a chance to see
those writes and ultimately cause gaps in the history.

We only do this when the continuous backup feature is enabled, and not
always, because the WAL replay can cause the history window to be
extended significantly (WAL is at 4GB * NumKVStore/NumCPU).

Change-Id: I51abb4852f4568bcef550ca628f8f73f4f727ec7
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/217963
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>