MB-63977: Tell the producer to send the purge-seqno and then store it
Add to the Consumer controls list "max_marker_version=2.2".
If the producer enables this then received markers can include the
purge-seqno.

The snapshotMarker receive path is now updated (the bulk of the
change) so that the "optional" purge-seqno is passed through the
callstack and processed in consumer.

With this patch merged, the replica will track the active purge-seqno
for all disk snapshots.

Change-Id: I9a517ec9be589008348ba75dc0165ad325771c78
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/218941
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Jim Walker <jim@couchbase.com>