MB-63972: Always close the MutationLog file
There is an error path in flush() where it sets the internal
variable "disabled" to true. This would cause the file to remain
open (and we would leak a file descriptor). The other paths call
disable(), but flush() cannot use that method as disable() would
call close() which in turn would call back into flush().

Change-Id: I14f828b2ec36dda85d8834e16204f5e92285a2ae
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/218311
Reviewed-by: Pavlos Georgiou <pavlos.georgiou@couchbase.com>
Tested-by: Trond Norbye <trond.norbye@couchbase.com>