MB-63979: Only reformat histograms when requested by cbstat
cbstat don't support dumping the JSON versions of the histograms
which means we have to convert it to the "old style" histogram
format where we encoded the buckets in the key.

We want to eventually remove cbstat so we don't want to spend
time adding support for generating histograms from the JSON
representation. Instead start returning the stat as JSON
unless the client identified itself as cbstat

Change-Id: Ic18dfdbe5ecf00dd5fdb60336edc92f81fbcf90b
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/218356
Tested-by: Trond Norbye <trond.norbye@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>