MB-63984: Do not count the listening sockets towards current_connections
The kv_current_connections stat description says:

  The current number of connections. This includes user, system and
daemon connections

However, that is not entirely accurate. That counter also includes the
number of server listening sockets. We've found that there are buggy
error paths in setting up listening sockets, which cause this metric
value to be incorrect. Instead of fixing the accounting, given that
is it probably unexpected for this metric to track that count in the
first place, it seems better to just stop tracking server sockets into
current_connections.

This change does so by splitting the function safe_close which was
responsible for decrementing the counter into a "server" and "client"
variant.

The test added is for the bug in the old behaviour, so it is technically
unnessary now that we don't increment that metric, but I've kept it for
completeness.

Change-Id: I8d58475f855275b73d5ba6cd7c938c946b77b491
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/218460
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Vesko Karaganev <vesko.karaganev@couchbase.com>