MB-62250: Use the correct bucket name and log paths correctly
Caught by test failures on Windows.

I was not using the correct bucket name config (it's couch_bucket) for
the bucket name, instead relying on EPEngine::getName() which includes a
colon character under unit tests and is not valid on Windows.

Additionally, the logging of the path was not showing correctly, as the
wchar_t* paths get printed as bytes sequences. Use generic_string() to
convert to char*.

Change-Id: Ie4ae1fa032e25ead3027374aa63b88c9212f6119
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/218585
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>