MB-60861: Make the ConnHandler logger instance prefix logs automatically
BucketLogger supports a setLogHeader which prefixes all log lines with
details about the DCP connection.

Change setLogHeader to setLogContext and also allow the caller to
provide a JSON object. The JSON object will be merged into the JSON log
context.

Some of this was done in ActiveStream and PassiveStream before.

Change-Id: I1ccb853d14ea200664a968e88870fb140793facd
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/218770
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>